requirements.txt
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
six==1.16.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
